Change double -> float.
This has no immediate speed impacts, but it will likely in the future.

+++ enc/bit_cost_inc.h
@@ -9,15 +9,15 @@
 
 #define HistogramType FN(Histogram)
 
-double FN(BrotliPopulationCost)(const HistogramType* histogram) {
-  static const double kOneSymbolHistogramCost = 12;
-  static const double kTwoSymbolHistogramCost = 20;
-  static const double kThreeSymbolHistogramCost = 28;
-  static const double kFourSymbolHistogramCost = 37;
+float FN(BrotliPopulationCost)(const HistogramType* histogram) {
+  static const float kOneSymbolHistogramCost = 12;
+  static const float kTwoSymbolHistogramCost = 20;
+  static const float kThreeSymbolHistogramCost = 28;
+  static const float kFourSymbolHistogramCost = 37;
   const size_t data_size = FN(HistogramDataSize)();
   int count = 0;
   size_t s[5];
-  double bits = 0.0;
+  float bits = 0.0;
   size_t i;
   if (histogram->total_count_ == 0) {
     return kOneSymbolHistogramCost;
@@ -33,7 +33,7 @@ double FN(BrotliPopulationCost)(const HistogramType* histogram) {
     return kOneSymbolHistogramCost;
   }
   if (count == 2) {
-    return (kTwoSymbolHistogramCost + (double)histogram->total_count_);
+    return (kTwoSymbolHistogramCost + (float)histogram->total_count_);
   }
   if (count == 3) {
     const uint32_t histo0 = histogram->data_[s[0]];
@@ -72,12 +72,12 @@ double FN(BrotliPopulationCost)(const HistogramType* histogram) {
        zero repeat code 17, but we don't use the non-zero repeat code 16. */
     size_t max_depth = 1;
     uint32_t depth_histo[BROTLI_CODE_LENGTH_CODES] = { 0 };
-    const double log2total = FastLog2(histogram->total_count_);
+    const float log2total = FastLog2(histogram->total_count_);
     for (i = 0; i < data_size;) {
       if (histogram->data_[i] > 0) {
         /* Compute -log2(P(symbol)) = -log2(count(symbol)/total_count) =
                                     = log2(total_count) - log2(count(symbol)) */
-        double log2p = log2total - FastLog2(histogram->data_[i]);
+        float log2p = log2total - FastLog2(histogram->data_[i]);
         /* Approximate the bit depth by round(-log2(P(symbol))) */
         size_t depth = (size_t)(log2p + 0.5);
         bits += histogram->data_[i] * log2p;
@@ -117,7 +117,7 @@ double FN(BrotliPopulationCost)(const HistogramType* histogram) {
       }
     }
     /* Add the estimated encoding cost of the code length code histogram. */
-    bits += (double)(18 + 2 * max_depth);
+    bits += (float)(18 + 2 * max_depth);
     /* Add the entropy of the code length code histogram. */
     bits += BitsEntropy(depth_histo, BROTLI_CODE_LENGTH_CODES);
   }
